Receive enodes over websockets and serve them at /static-nodes.json
Intention is to make `eth-netstats` a source of Auto-updated Enode List
to use with Nethermind of example.

How it works: The enode-reporter is a separate NodeJS app that is designed to run
besides Nethermind and query the Nethermind client via JsonRPC. Then it publishes enode and name
on the `enode` channel via Websockets connection with `eth-netstats`. Here we collect all
enodes (keyed on node id for deduplication purposes) and store them
in the format of `static-nodes.json`, one of the config
files for Nethermind.

The file is available at `/static-nodes.json`.